requirements.txt
fastapi
uvicorn
streamlit
groq
python-dotenv
python-multipart
pillow
requests
